requirements.txt
Django==4.1
djangorestframework==3.13.1
geopy==2.2.0
coverage==6.4.4
pylint==2.14.5


